onComplete refactor
